Crate thermite

Re-exports

pub use element::SimdElement;

pub use math::SimdVectorizedMath;

pub use math::*;

Modules

arch
Organized target-feature intrinsics

backends

element

math
Vectorized Math Library

rng

Macros

shuffle
Shuffles the elements in one or two vectors into a new vector given the indices provided.

Structs

BitMask
Small integer representation of a mask using individual bits

BranchfreeDivider
Divider without branching, useful for dynamic divisors.

Divider
Divider recommended for constant divisors.

Mask

VPtr

VectorBuffer
Aligned SIMD vector storage

Enums

SimdInstructionSet
Enum of supported instruction sets

Traits

Simd
SIMD Instruction set, contains all types

SimdBitwise
Defines bitwise operations on vectors

SimdCastTo
Describes casting to one SIMD vector type from another

SimdCasts
List of valid casts between SIMD types in an instruction set

SimdFloatVector
Floating point SIMD vectors

SimdFromBits
Transmutations from raw bits

SimdFromCast
Describes casting from one SIMD vector type to another

SimdIntVector
Integer SIMD vectors

SimdIntegerDivision
Specialized integer division by Dividers

SimdIntoBits
Transmutations into raw bits

SimdPointer
Guarantees the vector can be used as a pointer in VPtr

SimdShuffleIndices
Helper trait for constant vector shuffles

SimdSignedVector
Signed SIMD vector, with negative numbers

SimdUnsignedIntVector
Unsigned SIMD vector

SimdVector
Defines common operations on numeric vectors

SimdVectorBase
Basic shared vector interface

Type Definitions

AssociatedVector
Associated vector type for a scalar type

Vf32

Vf64

Vi32

Vi64

Visize

Vu32

Vu64

Vusize

Attribute Macros

dispatch
Generates monomorphized backend target_feature function calls to the annotated function or impl block.