Module thermite::arch::fma

Functions

_mm256_fmadd_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and add the intermediate result to packed elements in c.

_mm256_fmadd_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and add the intermediate result to packed elements in c.

_mm256_fmaddsub_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and alternatively add and subtract packed elements in c to/from the intermediate result.

_mm256_fmaddsub_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and alternatively add and subtract packed elements in c to/from the intermediate result.

_mm256_fmsub_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and subtract packed elements in c from the intermediate result.

_mm256_fmsub_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and subtract packed elements in c from the intermediate result.

_mm256_fmsubadd_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and alternatively subtract and add packed elements in c from/to the intermediate result.

_mm256_fmsubadd_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and alternatively subtract and add packed elements in c from/to the intermediate result.

_mm256_fnmadd_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and add the negated intermediate result to packed elements in c.

_mm256_fnmadd_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and add the negated intermediate result to packed elements in c.

_mm256_fnmsub_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and subtract packed elements in c from the negated intermediate result.

_mm256_fnmsub_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and subtract packed elements in c from the negated intermediate result.

_mm_fmadd_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and add the intermediate result to packed elements in c.

_mm_fmadd_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and add the intermediate result to packed elements in c.

_mm_fmadd_sd⚠	fma
Multiplies the lower double-precision (64-bit) floating-point elements in a and b, and add the intermediate result to the lower element in c. Stores the result in the lower element of the returned value, and copy the upper element from a to the upper elements of the result.

_mm_fmadd_ss⚠	fma
Multiplies the lower single-precision (32-bit) floating-point elements in a and b, and add the intermediate result to the lower element in c. Stores the result in the lower element of the returned value, and copy the 3 upper elements from a to the upper elements of the result.

_mm_fmaddsub_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and alternatively add and subtract packed elements in c to/from the intermediate result.

_mm_fmaddsub_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and alternatively add and subtract packed elements in c to/from the intermediate result.

_mm_fmsub_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and subtract packed elements in c from the intermediate result.

_mm_fmsub_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and subtract packed elements in c from the intermediate result.

_mm_fmsub_sd⚠	fma
Multiplies the lower double-precision (64-bit) floating-point elements in a and b, and subtract the lower element in c from the intermediate result. Store the result in the lower element of the returned value, and copy the upper element from a to the upper elements of the result.

_mm_fmsub_ss⚠	fma
Multiplies the lower single-precision (32-bit) floating-point elements in a and b, and subtract the lower element in c from the intermediate result. Store the result in the lower element of the returned value, and copy the 3 upper elements from a to the upper elements of the result.

_mm_fmsubadd_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and alternatively subtract and add packed elements in c from/to the intermediate result.

_mm_fmsubadd_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and alternatively subtract and add packed elements in c from/to the intermediate result.

_mm_fnmadd_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and add the negated intermediate result to packed elements in c.

_mm_fnmadd_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and add the negated intermediate result to packed elements in c.

_mm_fnmadd_sd⚠	fma
Multiplies the lower double-precision (64-bit) floating-point elements in a and b, and add the negated intermediate result to the lower element in c. Store the result in the lower element of the returned value, and copy the upper element from a to the upper elements of the result.

_mm_fnmadd_ss⚠	fma
Multiplies the lower single-precision (32-bit) floating-point elements in a and b, and add the negated intermediate result to the lower element in c. Store the result in the lower element of the returned value, and copy the 3 upper elements from a to the upper elements of the result.

_mm_fnmsub_pd⚠	fma
Multiplies packed double-precision (64-bit) floating-point elements in a and b, and subtract packed elements in c from the negated intermediate result.

_mm_fnmsub_ps⚠	fma
Multiplies packed single-precision (32-bit) floating-point elements in a and b, and subtract packed elements in c from the negated intermediate result.

_mm_fnmsub_sd⚠	fma
Multiplies the lower double-precision (64-bit) floating-point elements in a and b, and subtract packed elements in c from the negated intermediate result. Store the result in the lower element of the returned value, and copy the upper element from a to the upper elements of the result.

_mm_fnmsub_ss⚠	fma
Multiplies the lower single-precision (32-bit) floating-point elements in a and b, and subtract packed elements in c from the negated intermediate result. Store the result in the lower element of the returned value, and copy the 3 upper elements from a to the upper elements of the result.