Macro thermite::shuffle

macro_rules! shuffle {
    ($a:expr, $b:expr, [$($idx:expr),*]) => { ... };
    ($a:expr, [$($idx:expr),*]) => { ... };
}

Shuffles the elements in one or two vectors into a new vector given the indices provided.

Under the hood this generates an anonymous struct implementing SimdShuffleIndices, then calls SimdVectorBase::shuffle on a vector or two. Shuffles will be monomorphized to generate ideal code wherever possible.

The length of the indices array is not checked at compile-time, but will panic at runtime if incorrect or if any of the elements are out of bounds for the vectors.

You can use a crate such as no-panic to statically ensure all the indices are valid. no-panic checks at link-time, so branches pruned via dead-code removal will not contribute, allowing you to do things like:

match Vf32::NUM_ELEMENTS {
    4 => shuffle!(x, y, [6, 2, 1, 7]),
    8 => shuffle!(x, y, [5, 6, 10, 9, 2, 8, 6, 4]),
    _ => unimplemented!(),
}