Struct thermite::math::complex::Complex

pub struct Complex<S: Simd, V: SimdFloatVector<S>> {
    pub re: V,
    pub im: V,
    // some fields omitted
}

Fields
re: Vim: V
Implementations

impl<S: Simd, V: SimdFloatVector<S>> Complex<S, V>

pub fn new(re: V, im: V) -> Self

pub fn splat(re: V::Element, im: V::Element) -> Self

Creates a new Complex with all lanes of re and im set to the inputs, respectively.

pub fn real(re: V) -> Self

Create a new Complex

pub fn i() -> Self

Returns imaginary unit

pub fn one() -> Self

real(1)

pub fn zero() -> Self

real(0)

impl<S: Simd, V: SimdFloatVector<S>> Complex<S, V> where
V: SimdVectorizedMath<S>,

pub fn norm_sqr(self) -> V

Returns the square of the norm

pub fn norm(self) -> V

Calculate |self|

pub fn scale(self, t: V) -> Self

Multiplies self by the scalar t.

pub fn unscale(self, t: V) -> Self

Divides self by the scalar t.

pub fn conj(self) -> Self

Returns the complex conjugate. i.e. re - i im

pub fn inv(self) -> Self

Returns 1/self

pub fn mul_add(self, m: Self, a: Self) -> Self

pub fn l1_norm(self) -> V

Returns the L1 norm |re| + |im| -- the Manhattan distance from the origin.

pub fn arg(self) -> V

Calculate the principal Arg of self.

pub fn to_polar(self) -> (V, V)

Convert to polar form (r, theta), such that self = r * exp(i * theta)

pub fn from_polar(r: V, theta: V) -> Self

Convert a polar representation into a complex number.

pub fn exp(self) -> Self

Computes e^(self), where e is the base of the natural logarithm.

pub fn ln(self) -> Self

Computes the principal value of natural logarithm of self.

This function has one branch cut:

(-∞, 0], continuous from above.

The branch satisfies -π ≤ arg(ln(z)) ≤ π.

pub fn sqrt(self) -> Self

pub fn cbrt(self) -> Self

pub fn powf(self, exp: V) -> Self

Raises self to a floating point power.

pub fn log(self, base: V) -> Self

Returns the logarithm of self with respect to an arbitrary base.

pub fn powc(self, exp: Self) -> Self

Raises self to a complex power.

pub fn expf(self, base: V) -> Self

Raises a floating point number to the complex power self.

pub fn sin(self) -> Self

Computes the sine of self.

pub fn cos(self) -> Self

Computes the cosine of self.

pub fn tan(self) -> Self

Computes the tangent of self.

pub fn asin(self) -> Self

Computes the principal value of the inverse sine of self.

This function has two branch cuts:

(-∞, -1), continuous from above.

(1, ∞), continuous from below.

The branch satisfies -π/2 ≤ Re(asin(z)) ≤ π/2.

pub fn acos(self) -> Self

Computes the principal value of the inverse cosine of self.

This function has two branch cuts:

(-∞, -1), continuous from above.

(1, ∞), continuous from below.

The branch satisfies 0 ≤ Re(acos(z)) ≤ π.

pub fn atan(self) -> Self

Computes the principal value of the inverse tangent of self.

This function has two branch cuts:

(-∞i, -i], continuous from the left.

[i, ∞i), continuous from the right.

The branch satisfies -π/2 ≤ Re(atan(z)) ≤ π/2.

pub fn sinh(self) -> Self

Computes the hyperbolic sine of self.

pub fn cosh(self) -> Self

Computes the hyperbolic cosine of self.

pub fn tanh(self) -> Self

Computes the hyperbolic tangent of self.

pub fn asinh(self) -> Self

Computes the principal value of inverse hyperbolic sine of self.

This function has two branch cuts:

(-∞i, -i), continuous from the left.

(i, ∞i), continuous from the right.

The branch satisfies -π/2 ≤ Im(asinh(z)) ≤ π/2.

pub fn acosh(self) -> Self

Computes the principal value of inverse hyperbolic cosine of self.

This function has one branch cut:

(-∞, 1), continuous from above.

The branch satisfies -π ≤ Im(acosh(z)) ≤ π and 0 ≤ Re(acosh(z)) < ∞.

pub fn atanh(self) -> Self

Computes the principal value of inverse hyperbolic tangent of self.

This function has two branch cuts:

(-∞, -1], continuous from above.

[1, ∞), continuous from below.

The branch satisfies -π/2 ≤ Im(atanh(z)) ≤ π/2.

pub fn finv(self) -> Self

Returns 1/self using floating-point operations.

This may be more accurate than the generic self.inv() in cases where self.norm_sqr() would overflow to ∞ or underflow to 0.

pub fn fdiv(self, rhs: Self) -> Self

Returns self/other using floating-point operations.

This may be more accurate than the generic Div implementation in cases where other.norm_sqr() would overflow to ∞ or underflow to 0.