Module thermite::math::policies

Execution Policies (precision, performance, etc.)

To define a custom policy:

pub struct MyPolicy;

impl Policy for MyPolicy {
    const POLICY: Parameters = Parameters {
        check_overflow: false,
        unroll_loops: false,
        precision: PrecisionPolicy::Average,
        avoid_branching: true,
        max_series_iterations: 10000,
    };
}

let y = x.cbrt_p::<MyPolicy>();

Structs

ExtraPrecision
Policy adapter that increases the precision requires by one level, e.g.: Worst -> Average, Average -> Best

Performance
Optimize for performance, ideally without losing precision.

Precision
Optimize for precision, at the cost of performance if necessary.

Reference
Calculates a reference value for operations where possible, which can be very expensive.

Size
Optimize for code size, avoids hard-coded equations or loop unrolling.

UltraPerformance
Optimize for performance at the cost of precision and safety (doesn’t handle special cases such as NaNs or overflow).