Struct thermite::math::policies::Parameters

pub struct Parameters {
    pub check_overflow: bool,
    pub unroll_loops: bool,
    pub extra_precision: bool,
    pub avoid_branching: bool,
}

Customizable Policy Parameters

Fields
check_overflow: bool

If true, methods will check for infinity/NaN/invalid domain issues and give a well-formed standard result.

If false, all of that work is avoided, and the result is undefined in those cases. Garbage in, garbage out.

However, those checks can be expensive.

unroll_loops: bool

If true, unrolled and optimized versions of some algorithms will be used. These can be much faster than the linear variants. If code size is important, this will improve codegen when used with opt-level=z

extra_precision: bool

If true, extra care is taken to improve precision of results, often at the cost of some performance.

avoid_branching: bool

If true, methods will not try to avoid extra work by branching. Some of the internal branches are expensive, but branchless may be desired in some cases, such as minimizing code size. However, note that some branches are required for improving precision. Enabling extra_precision will re-enable those.

Implementations

impl Parameters

pub const fn avoid_precision_branches(self) -> bool

Returns true if the policy says to avoid branches at the cost of precision