Module thermite::math::poly

Optimized fixed-degree polynomial evaluation

All of these polynomials use Estrin's scheme to reduce the dependency chain length and encourage instruction-level parallelism, which has the potential to improve performance despite the powers of X being required upfront.

Powers of x are required, rather than computed internally, so they could be reused between multiple polynomials.

Unless you are micro-optimizing, it's recommended to use SimdVectorizedMath::poly or poly_f

Functions

poly_1

poly_2

poly_3

poly_4

poly_5

poly_6

poly_7

poly_8

poly_9

poly_10

poly_11

poly_12

poly_13

poly_14

poly_15

poly_30