Trait thermite::math::SimdVectorizedMathPolicied

pub trait SimdVectorizedMathPolicied<S: Simd>: SimdFloatVector<S> + SimdFloatVectorConsts<S> {

    fn scale_p<P: Policy>(
        self, 
        in_min: Self, 
        in_max: Self, 
        out_min: Self, 
        out_max: Self
    ) -> Self;

    fn lerp_p<P: Policy>(self, a: Self, b: Self) -> Self;

    fn fmod_p<P: Policy>(self, y: Self) -> Self;

    fn hypot_p<P: Policy>(self, y: Self) -> Self;

    fn poly_p<P: Policy>(self, coefficients: &[Self::Element]) -> Self;

    fn poly_rational_p<P: Policy>(
        self, 
        numerator_coefficients: &[Self::Element], 
        denominator_coefficients: &[Self::Element]
    ) -> Self;

    fn poly_f_p<P: Policy, F>(self, n: usize, f: F) -> Self
    where
        F: FnMut(usize) -> Self;

    fn summation_f_p<P: Policy, F>(
        start: isize, 
        end: isize, 
        f: F
    ) -> Result<Self, Self>
    where
        F: FnMut(isize) -> Self;

    fn product_f_p<P: Policy, F>(
        start: isize, 
        end: isize, 
        f: F
    ) -> Result<Self, Self>
    where
        F: FnMut(isize) -> Self;

    fn sum_f_p<P: Policy, F>(n: usize, f: F) -> Self
    where
        F: FnMut(usize) -> Self;

    fn sin_p<P: Policy>(self) -> Self;

    fn cos_p<P: Policy>(self) -> Self;

    fn tan_p<P: Policy>(self) -> Self;

    fn sin_cos_p<P: Policy>(self) -> (Self, Self);

    fn sinh_p<P: Policy>(self) -> Self;

    fn cosh_p<P: Policy>(self) -> Self;

    fn tanh_p<P: Policy>(self) -> Self;

    fn asinh_p<P: Policy>(self) -> Self;

    fn acosh_p<P: Policy>(self) -> Self;

    fn atanh_p<P: Policy>(self) -> Self;

    fn asin_p<P: Policy>(self) -> Self;

    fn acos_p<P: Policy>(self) -> Self;

    fn atan_p<P: Policy>(self) -> Self;

    fn atan2_p<P: Policy>(self, x: Self) -> Self;

    fn exp_p<P: Policy>(self) -> Self;

    fn exph_p<P: Policy>(self) -> Self;

    fn exp2_p<P: Policy>(self) -> Self;

    fn exp10_p<P: Policy>(self) -> Self;

    fn exp_m1_p<P: Policy>(self) -> Self;

    fn cbrt_p<P: Policy>(self) -> Self;

    fn invsqrt_p<P: Policy>(self) -> Self;

    fn reciprocal_p<P: Policy>(self) -> Self;

    fn powf_p<P: Policy>(self, e: Self) -> Self;

    fn powiv_p<P: Policy>(self, e: S::Vi32) -> Self;

    fn powi_p<P: Policy>(self, e: i32) -> Self;

    fn ln_p<P: Policy>(self) -> Self;

    fn ln_1p_p<P: Policy>(self) -> Self;

    fn log2_p<P: Policy>(self) -> Self;

    fn log10_p<P: Policy>(self) -> Self;

    fn erf_p<P: Policy>(self) -> Self;

    fn erfinv_p<P: Policy>(self) -> Self;

    fn next_float_p<P: Policy>(self) -> Self;

    fn prev_float_p<P: Policy>(self) -> Self;

    fn smoothstep_p<P: Policy>(self) -> Self;

    fn smootherstep_p<P: Policy>(self) -> Self;

    fn smootheststep_p<P: Policy>(self) -> Self;

}

Set of vectorized special functions allowing specific execution policies

Please refer to the documentation of SimdVectorizedMath for function reference and policies for an overview of available execution policies.

Required methods

fn scale_p<P: Policy>(
self,
in_min: Self,
in_max: Self,
out_min: Self,
out_max: Self
) -> Self

fn lerp_p<P: Policy>(self, a: Self, b: Self) -> Self

fn fmod_p<P: Policy>(self, y: Self) -> Self

fn hypot_p<P: Policy>(self, y: Self) -> Self

fn poly_p<P: Policy>(self, coefficients: &[Self::Element]) -> Self

fn poly_rational_p<P: Policy>(
self,
numerator_coefficients: &[Self::Element],
denominator_coefficients: &[Self::Element]
) -> Self

fn poly_f_p<P: Policy, F>(self, n: usize, f: F) -> Self where
F: FnMut(usize) -> Self,

fn summation_f_p<P: Policy, F>(
start: isize,
end: isize,
f: F
) -> Result<Self, Self> where
F: FnMut(isize) -> Self,

fn product_f_p<P: Policy, F>(
start: isize,
end: isize,
f: F
) -> Result<Self, Self> where
F: FnMut(isize) -> Self,

fn sum_f_p<P: Policy, F>(n: usize, f: F) -> Self where
F: FnMut(usize) -> Self,

fn sin_p<P: Policy>(self) -> Self

fn cos_p<P: Policy>(self) -> Self

fn tan_p<P: Policy>(self) -> Self

fn sin_cos_p<P: Policy>(self) -> (Self, Self)

fn sinh_p<P: Policy>(self) -> Self

fn cosh_p<P: Policy>(self) -> Self

fn tanh_p<P: Policy>(self) -> Self

fn asinh_p<P: Policy>(self) -> Self

fn acosh_p<P: Policy>(self) -> Self

fn atanh_p<P: Policy>(self) -> Self

fn asin_p<P: Policy>(self) -> Self

fn acos_p<P: Policy>(self) -> Self

fn atan_p<P: Policy>(self) -> Self

fn atan2_p<P: Policy>(self, x: Self) -> Self

fn exp_p<P: Policy>(self) -> Self

fn exph_p<P: Policy>(self) -> Self

fn exp2_p<P: Policy>(self) -> Self

fn exp10_p<P: Policy>(self) -> Self

fn exp_m1_p<P: Policy>(self) -> Self

fn cbrt_p<P: Policy>(self) -> Self

fn invsqrt_p<P: Policy>(self) -> Self

fn reciprocal_p<P: Policy>(self) -> Self

fn powf_p<P: Policy>(self, e: Self) -> Self

fn powiv_p<P: Policy>(self, e: S::Vi32) -> Self

fn powi_p<P: Policy>(self, e: i32) -> Self

fn ln_p<P: Policy>(self) -> Self

fn ln_1p_p<P: Policy>(self) -> Self

fn log2_p<P: Policy>(self) -> Self

fn log10_p<P: Policy>(self) -> Self

fn erf_p<P: Policy>(self) -> Self

fn erfinv_p<P: Policy>(self) -> Self

fn next_float_p<P: Policy>(self) -> Self

fn prev_float_p<P: Policy>(self) -> Self

fn smoothstep_p<P: Policy>(self) -> Self

fn smootherstep_p<P: Policy>(self) -> Self

fn smootheststep_p<P: Policy>(self) -> Self

Loading content...