Struct thermite::BitMask

#[repr(transparent)]pub struct BitMask<S: Simd, V> { /* fields omitted */ }

Small integer representation of a mask using individual bits

Implementations

impl<S: Simd, V> BitMask<S, V> where
V: SimdBitwise<S>,

pub fn truthy() -> Self

Create a bitmask for this vector type of all true bits

pub fn falsey() -> Self

Create a bitmask with all false bits

pub fn raw(self) -> u16

pub fn all(self) -> bool

Return true if all lanes for this vector type are set to true

pub fn any(self) -> bool

Return true if any lane for this vector type is set to true

pub fn none(self) -> bool

Return true if no lanes for this vector type are set to true

pub fn count(self) -> u32

Count the number of set bits