Struct thermite::BranchfreeDivider

#[repr(transparent)]
pub struct BranchfreeDivider<T>(_);

Divider without branching, useful for dynamic divisors.

However, when used with constant input, this may perform extra unnecessary work that could be removed in the branching Divider

Implementations

impl BranchfreeDivider<u8>

pub const fn u8(d: u8) -> Self

See docs for BranchfreeDivider and Divider

impl BranchfreeDivider<u16>

pub const fn u16(d: u16) -> Self

See docs for BranchfreeDivider and Divider

impl BranchfreeDivider<u32>

pub const fn u32(d: u32) -> Self

See docs for BranchfreeDivider and Divider

impl BranchfreeDivider<u64>

pub const fn u64(d: u64) -> Self

See docs for BranchfreeDivider and Divider

impl BranchfreeDivider<i8>

pub const fn i8(d: i8) -> Self

See docs for BranchfreeDivider and Divider

impl BranchfreeDivider<i16>

pub const fn i16(d: i16) -> Self

See docs for BranchfreeDivider and Divider

impl BranchfreeDivider<i32>

pub const fn i32(d: i32) -> Self

See docs for BranchfreeDivider and Divider

impl BranchfreeDivider<i64>

pub const fn i64(d: i64) -> Self

See docs for BranchfreeDivider and Divider