Struct thermite::Divider

#[repr(C, packed)]pub struct Divider<T> { /* fields omitted */ }

Divider recommended for constant divisors.

When using constant divisors, divisions using this can remove extra branches and generate ideal integer division code.

However, when used with dynamic input, the extra branches can be expensive, therefore it is recommended to use the branchfree alternative for dynamic divisors.

Implementations

impl<T> Divider<T>

pub fn multiplier(&self) -> T

pub fn shift(&self) -> u8

impl Divider<u8>

pub const fn u8(d: u8) -> Self

See docs for Divider

pub const fn u8_branchfree(d: u8) -> BranchfreeDivider<u8>

See docs for BranchfreeDivider and Divider

impl Divider<u16>

pub const fn u16(d: u16) -> Self

See docs for Divider

pub const fn u16_branchfree(d: u16) -> BranchfreeDivider<u16>

See docs for BranchfreeDivider and Divider

impl Divider<u32>

pub const fn u32(d: u32) -> Self

See docs for Divider

pub const fn u32_branchfree(d: u32) -> BranchfreeDivider<u32>

See docs for BranchfreeDivider and Divider

impl Divider<u64>

pub const fn u64(d: u64) -> Self

See docs for Divider

pub const fn u64_branchfree(d: u64) -> BranchfreeDivider<u64>

See docs for BranchfreeDivider and Divider