Struct thermite::Mask

#[repr(transparent)]
pub struct Mask<S: Simd, V>(_, _);

Implementations

impl<S: Simd + ?Sized, V> Mask<S, V>

pub fn value(self) -> V

impl<S: Simd + ?Sized, V> Mask<S, V> where
V: SimdVectorBase<S>,

pub fn truthy() -> Self

Mask vector containing all true/non-zero lanes.

pub fn falsey() -> Self

Mask vector containing all zero/false lanes.

impl<S: Simd + ?Sized, V> Mask<S, V> where
V: SimdVector<S>,

pub fn from_value(v: V) -> Self

Creates a mask from a value. Any non-zero lanes are considered truthy.

pub fn cast_to<U: SimdFromCast<S, V>>(self) -> Mask<S, U>

pub const FULL_BITMASK: BitMask<S, V>

Bitmask indicating all truthy values for each lane

pub fn and_not(self, other: Self) -> Self

Computes !self & other for each lane of the mask.

pub fn bitmask(self) -> BitMask<S, V>

Returns a bitmask that can be trivially evaluated to all/any/none/etc.

The can be useful for reusing a mask value for branching

pub fn from_bitmask(bitmask: u16) -> Self

Creates a wide SIMD mask from a single-bit bitmask

pub fn all(self) -> bool

Returns true if all lanes are truthy

NOTE: If you wish to use the same mask for multiply calls to all/any/none, consider using bitmask() instead to precompute the bitmask and reuse its methods.

pub fn any(self) -> bool

Returns true if any lanes are truthy

NOTE: If you wish to use the same mask for multiply calls to all/any/none, consider using bitmask() instead to precompute the bitmask and reuse its methods.

pub fn none(self) -> bool

Returns true if all lanes are falsey

NOTE: If you wish to use the same mask for multiply calls to all/any/none, consider using bitmask() instead to precompute the bitmask and reuse its methods.

pub fn count(self) -> u32

Counts the number of truthy lanes

pub fn select<U>(self, t: U, f: U) -> U where
V: SimdCastTo<S, U>,
U: SimdVector<S>,

For each lane, selects from t if the mask lane is truthy, or f is falsey

pub fn swap<U>(self, a: &mut U, b: &mut U) where
V: SimdCastTo<S, U>,
U: SimdVector<S>,

For each lane, if the mask lane is truthy then swap the corresponding lanes in a and b