Struct thermite::VectorBuffer

#[repr(transparent)]pub struct VectorBuffer<S: Simd, V: SimdVectorBase<S>> { /* fields omitted */ }

Aligned SIMD vector storage

Implementations

impl<S: Simd, V: SimdVectorBase<S>> VectorBuffer<S, V>

pub fn alloc(count: usize) -> Self

Allocates a new SIMD-aligned element buffer and zeroes the elements.

Due to the alignment, it will round up the number of elements to the nearest multiple of V::NUM_ELEMENTS, making the "wasted" space visible.

pub fn iter_vectors<'a>(&'a self) -> SimdSliceIter<'a, S, V>

pub fn iter_vectors_mut<'a>(&'a mut self) -> AlignedMutIter<'a, S, V>

pub fn gather(&self, indices: S::Vu32) -> V where
V: SimdVector<S>,

Gathers values from the buffer using more efficient instructions where possible

pub fn fill(&mut self, value: V)

Fills the buffer with vectors using aligned stores

pub fn len(&self) -> usize

pub fn len_vectors(&self) -> usize

pub fn as_slice(&self) -> &[V::Element]

pub fn as_vector_slice(&self) -> &[V]

pub fn as_mut_slice(&mut self) -> &mut [V::Element]

pub fn as_mut_vector_slice(&mut self) -> &mut [V]

pub fn load_vector(&self, vector_index: usize) -> V

pub fn store_vector(&mut self, vector_index: usize, value: V)