Trait thermite::SimdCastTo

pub trait SimdCastTo<S: Simd, TO>: Sized {
    fn cast(self) -> TO;

    fn cast_mask(mask: Mask<S, Self>) -> Mask<S, TO>;
}

Describes casting to one SIMD vector type from another

Required methods

fn cast(self) -> TO

Casts one vector to another, performing proper numeric conversions on each element.

This is equivalent to the as keyword in Rust, but for SIMD vectors.

fn cast_mask(mask: Mask<S, Self>) -> Mask<S, TO>

Casts one mask to another, not caring about the value types, but rather expanding or truncating the mask bits as efficiently as possible.

Loading content...