Trait thermite::SimdIntVector

pub trait SimdIntVector<S: Simd + ?Sized>: SimdVector<S> + Eq {
    fn saturating_add(self, rhs: Self) -> Self;

    fn saturating_sub(self, rhs: Self) -> Self;

    fn wrapping_sum(self) -> Self::Element;

    fn wrapping_product(self) -> Self::Element;

    fn rolv(self, cnt: S::Vu32) -> Self;

    fn rorv(self, cnt: S::Vu32) -> Self;

    fn reverse_bits(self) -> Self;

    fn count_ones(self) -> Self;

    fn leading_zeros(self) -> Self;

    fn trailing_zeros(self) -> Self;

    fn rol(self, cnt: u32) -> Self { ... }

    fn ror(self, cnt: u32) -> Self { ... }

    fn count_zeros(self) -> Self { ... }

    fn leading_ones(self) -> Self { ... }

    fn trailing_ones(self) -> Self { ... }
}

Integer SIMD vectors

Required methods

fn saturating_add(self, rhs: Self) -> Self

Saturating addition, will not wrap

fn saturating_sub(self, rhs: Self) -> Self

Saturating subtraction, will not wrap

fn wrapping_sum(self) -> Self::Element

Sum all lanes together, wrapping the result if it can't fit in T

fn wrapping_product(self) -> Self::Element

Multiply all lanes together, wrapping the result if it can't fit in T

fn rolv(self, cnt: S::Vu32) -> Self

Rotates the bits in each lane to the left (towards HSB) by the number of bits specified in the corresponding lane of cnt

fn rorv(self, cnt: S::Vu32) -> Self

Rotates the bits in each lane to the right (towards LSB) by the number of bits specified in the corresponding lane of cnt

fn reverse_bits(self) -> Self

Reverses the bits of each lane in the vector.

fn count_ones(self) -> Self

Counts the number of 1 bits in each lane of the vector.

fn leading_zeros(self) -> Self

Counts the number of leading zeros in each lane of the vector.

fn trailing_zeros(self) -> Self

Counts the number of trailing zeros in each lane of the vector.

Loading content...
Provided methods

fn rol(self, cnt: u32) -> Self

Rotates the bits in each lane to the left (towards HSB) by the number of bits specified in cnt

fn ror(self, cnt: u32) -> Self

Rotates the bits in each lane to the right (towards LSB) by the number of bits specified in cnt

fn count_zeros(self) -> Self

Counts the number of 0 bits in each lane of the vector.

fn leading_ones(self) -> Self

Counts the number of leading ones in each lane of the vector.

fn trailing_ones(self) -> Self

Counts the number of trailing ones in each lane of the vector.

Loading content...