Trait thermite::SimdUnsignedIntVector

pub trait SimdUnsignedIntVector<S: Simd + ?Sized>: SimdIntVector<S> {
    fn next_power_of_two_m1(self) -> Self;

    fn log2p1(self) -> Self { ... }

    fn is_power_of_two(self) -> Mask<S, Self> { ... }
}

Unsigned SIMD vector

Required methods

fn next_power_of_two_m1(self) -> Self

Returns next_power_of_two(x) - 1

Loading content...
Provided methods

fn log2p1(self) -> Self

Returns floor(log2(x)) + 1

fn is_power_of_two(self) -> Mask<S, Self>

Returns a mask wherein if a lane was a power of two, the corresponding mask lane will be truthy

Loading content...