Trait thermite::SimdVectorBase

pub trait SimdVectorBase<S: Simd + ?Sized>: Sized + Copy + Debug + Default + Send + Sync {
    type Element: SimdElement;

    const ELEMENT_SIZE: usize;
    const NUM_ELEMENTS: usize;
    const ALIGNMENT: usize;

    fn splat(value: Self::Element) -> Self;

    unsafe fn shuffle_unchecked<INDICES: SimdShuffleIndices>(
        self, 
        b: Self, 
        indices: INDICES
    ) -> Self;

    unsafe fn extract_unchecked(self, index: usize) -> Self::Element;

    unsafe fn replace_unchecked(
        self, 
        index: usize, 
        value: Self::Element
    ) -> Self;

    unsafe fn undefined() -> Self { ... }

    fn splat_any(value: impl Into<Self::Element>) -> Self { ... }

    fn splat_as<E>(value: E) -> Self
    where
        Self::Element: CastFrom<E>,
    { ... }

    fn shuffle<INDICES: SimdShuffleIndices>(
        self, 
        b: Self, 
        indices: INDICES
    ) -> Self { ... }

    fn shuffle_dyn(self, b: Self, indices: &[usize]) -> Self { ... }

    unsafe fn shuffle_dyn_unchecked(self, b: Self, indices: &[usize]) -> Self { ... }

    fn alloc(count: usize) -> VectorBuffer<S, Self> { ... }

    fn extract(self, index: usize) -> Self::Element { ... }

    fn replace(self, index: usize, value: Self::Element) -> Self { ... }

    fn load_aligned(src: &[Self::Element]) -> Self { ... }

    fn load_unaligned(src: &[Self::Element]) -> Self { ... }

    fn store_aligned(self, dst: &mut [Self::Element]) { ... }

    fn store_unaligned(self, dst: &mut [Self::Element]) { ... }

    unsafe fn load_aligned_unchecked(src: *const Self::Element) -> Self { ... }

    unsafe fn store_aligned_unchecked(self, dst: *mut Self::Element) { ... }

    unsafe fn load_unaligned_unchecked(src: *const Self::Element) -> Self { ... }

    unsafe fn store_unaligned_unchecked(self, dst: *mut Self::Element) { ... }

    unsafe fn gather_unchecked(
        src: *const Self::Element, 
        indices: S::Vi32
    ) -> Self { ... }

    unsafe fn scatter_unchecked(self, dst: *mut Self::Element, indices: S::Vi32) { ... }

    unsafe fn gather_masked_unchecked(
        base_ptr: *const Self::Element, 
        indices: S::Vi32, 
        mask: Mask<S, Self>, 
        default: Self
    ) -> Self { ... }

    unsafe fn scatter_masked_unchecked(
        self, 
        base_ptr: *mut Self::Element, 
        indices: S::Vi32, 
        mask: Mask<S, Self>
    ) { ... }
}

Basic shared vector interface

Associated Types

type Element: SimdElement

Loading content...
Associated Constants

const ELEMENT_SIZE: usize

Size of element type in bytes

const NUM_ELEMENTS: usize

const ALIGNMENT: usize

Loading content...
Required methods

fn splat(value: Self::Element) -> Self

Creates a new vector with all lanes set to the given value

unsafe fn shuffle_unchecked<INDICES: SimdShuffleIndices>(
self,
b: Self,
indices: INDICES
) -> Self

Shuffles between two vectors based on the static indices provided in INDICES

See the shuffle! macro for more information.

NOTE: Calling this with invalid indices (incorrect length or out-of-bounds values) will result in undefined behavior.

unsafe fn extract_unchecked(self, index: usize) -> Self::Element

Extracts an element at the given lane index.

WARNING: Will result in undefined behavior if the index is not less than NUM_ELEMENTS

unsafe fn replace_unchecked(self, index: usize, value: Self::Element) -> Self

Returns a new vector with the given value at the given lane index.

WARNING: Will result in undefined behavior if the index is not less than NUM_ELEMENTS

Loading content...
Provided methods

unsafe fn undefined() -> Self

Returns a vector containing possibly undefined or uninitialized data

fn splat_any(value: impl Into<Self::Element>) -> Self

Same as splat, but is more convenient for initializing with data that can be converted into the element type.

fn splat_as<E>(value: E) -> Self where
Self::Element: CastFrom<E>,

Splats a value by casting to the element type via value as Element.

This is limited to primitive numeric types.

fn shuffle<INDICES: SimdShuffleIndices>(self, b: Self, indices: INDICES) -> Self

Shuffles between two vectors based on the static indices provided in INDICES

See the shuffle! macro for more information.

NOTE: This method will panic if the indices are the incorrect length or out of bounds.

fn shuffle_dyn(self, b: Self, indices: &[usize]) -> Self

Shuffles between two vectors based on the dynamic indices provided.

This differs from the shuffle! macro and Self::shuffle methods in that these indices can indeed be dynamic, at the cost of performance.

NOTE: This method will panic if the indices are the incorrect length or out of bounds.

unsafe fn shuffle_dyn_unchecked(self, b: Self, indices: &[usize]) -> Self

Like Self::shuffle_dyn, but does not check for valid indices or input length.

NOTE: Calling this with invalid indices (incorrect length or out-of-bounds values) will result in undefined behavior.

fn alloc(count: usize) -> VectorBuffer<S, Self>

fn extract(self, index: usize) -> Self::Element

Extracts an element at the given lane index.

WARNING: Will panic if the index is not less than NUM_ELEMENTS

fn replace(self, index: usize, value: Self::Element) -> Self

Returns a new vector with the given value at the given lane index.

WARNING: Will panic if the index if not less than NUM_ELEMENTS

fn load_aligned(src: &[Self::Element]) -> Self

Loads a vector from a slice that has an alignment of at least Self::ALIGNMENT

WARNING: Will panic if the slice is not properly aligned or is not long enough.

fn load_unaligned(src: &[Self::Element]) -> Self

Loads a vector from a slice

WARNING: Will panic if the slice is not long enough.

fn store_aligned(self, dst: &mut [Self::Element])

Stores a vector into a slice with an alignment of at least Self::ALIGNMENT

WARNING: Will panic if the target slice is not properly aligned or is not long enough.

fn store_unaligned(self, dst: &mut [Self::Element])

Stores a vector into a slice.

WARNING: Will panic if the slice is not long enough.

unsafe fn load_aligned_unchecked(src: *const Self::Element) -> Self

Loads a vector from the given aligned address.

WARNING: Will cause undefined behavior if the pointer is not properly aligned or does not point to a valid address range.

unsafe fn store_aligned_unchecked(self, dst: *mut Self::Element)

Stores a vector to the given aligned address.

WARNING: Will cause undefined behavior if the pointer is not properly aligned or does not point to a valid address range.

unsafe fn load_unaligned_unchecked(src: *const Self::Element) -> Self

Loads a vector from a given address (does not have to be aligned).

WARNING: Will cause undefined behavior if the pointer does not point to a valid address range.

unsafe fn store_unaligned_unchecked(self, dst: *mut Self::Element)

Stores a vector to a given address (does not have to be aligned).

WARNING: Will cause undefined behavior if the pointer does not point to a valid address range.

unsafe fn gather_unchecked(src: *const Self::Element, indices: S::Vi32) -> Self

unsafe fn scatter_unchecked(self, dst: *mut Self::Element, indices: S::Vi32)

unsafe fn gather_masked_unchecked(
base_ptr: *const Self::Element,
indices: S::Vi32,
mask: Mask<S, Self>,
default: Self
) -> Self

Like Self::gather, but individual lanes are loaded based on the corresponding lane of the mask. If the mask lane is truthy, the source lane is loaded, otherwise it's given the lane value from default.

Lanes with a falsey mask value do not load, and does not cause undefined behavior if the source address is invalid for that lane.

unsafe fn scatter_masked_unchecked(
self,
base_ptr: *mut Self::Element,
indices: S::Vi32,
mask: Mask<S, Self>
)

Like self.scatter(), but individual lanes are stored based on the corresponding lane of the mask. If the mask lane is truthy, the destination lane is written to, otherwise it is a no-op.

Lanes with a falsey mask value are not written to, and does not cause undefined behavior if the destination address is invalid for that lane.

Loading content...