Trait thermite::SimdVectorizedMath

pub trait SimdVectorizedMath<S: Simd>: SimdFloatVector<S> {
    fn scale(
        self, 
        in_min: Self, 
        in_max: Self, 
        out_min: Self, 
        out_max: Self
    ) -> Self;

    fn lerp(self, a: Self, b: Self) -> Self;

    fn fmod(self, y: Self) -> Self;

    fn hypot(self, y: Self) -> Self;

    fn sin(self) -> Self;

    fn cos(self) -> Self;

    fn tan(self) -> Self;

    fn sin_cos(self) -> (Self, Self);

    fn sinh(self) -> Self;

    fn cosh(self) -> Self;

    fn tanh(self) -> Self;

    fn asinh(self) -> Self;

    fn acosh(self) -> Self;

    fn atanh(self) -> Self;

    fn asin(self) -> Self;

    fn acos(self) -> Self;

    fn atan(self) -> Self;

    fn atan2(self, x: Self) -> Self;

    fn exp(self) -> Self;

    fn exph(self) -> Self;

    fn exp2(self) -> Self;

    fn exp10(self) -> Self;

    fn exp_m1(self) -> Self;

    fn powf(self, e: Self) -> Self;

    fn powiv(self, e: S::Vi32) -> Self;

    fn powi(self, e: i32) -> Self;

    fn ln(self) -> Self;

    fn ln_1p(self) -> Self;

    fn log2(self) -> Self;

    fn log10(self) -> Self;

    fn erf(self) -> Self;

    fn erfinv(self) -> Self;

    fn next_float(self) -> Self;

    fn prev_float(self) -> Self;

    fn smoothstep(self) -> Self;

    fn smootherstep(self) -> Self;

    fn smootheststep(self) -> Self;

    fn hermite(self, n: u32) -> Self;

    fn hermitev(self, n: S::Vu32) -> Self;

    fn jacobi(self, alpha: Self, beta: Self, n: u32) -> Self;
}

Set of vectorized special functions optimized for both single and double precision

Required methods

fn scale(self, in_min: Self, in_max: Self, out_min: Self, out_max: Self) -> Self

Scales values between in_min and in_max, to between out_min and out_max

fn lerp(self, a: Self, b: Self) -> Self

Linearly interpolates between a and b using self

Equivalent to (1 - t) * a + t * b, but uses fused multiply-add operations to improve performance while maintaining precision

fn fmod(self, y: Self) -> Self

Returns the floating-point remainder of self / y (rounded towards zero)

fn hypot(self, y: Self) -> Self

Computes sqrt(x * x + y * y) for each element of the vector, but can be more precise with values around zero.

NOTE: This is not higher-performance than the naive version, only slightly more precise.

fn sin(self) -> Self

Computes the sine of a vector.

fn cos(self) -> Self

Computes the cosine of a vector.

fn tan(self) -> Self

Computes the tangent of a vector.

fn sin_cos(self) -> (Self, Self)

Computes both the sine and cosine of a vector together faster than they would be computed separately.

If you need both, use this.

fn sinh(self) -> Self

Computes the hyperbolic-sine of a vector.

fn cosh(self) -> Self

Computes the hyperbolic-cosine of a vector.

fn tanh(self) -> Self

Computes the hyperbolic-tangent of a vector.

fn asinh(self) -> Self

Computes the hyperbolic-arcsine of a vector.

fn acosh(self) -> Self

Computes the hyperbolic-arccosine of a vector.

fn atanh(self) -> Self

Computes the hyperbolic-arctangent of a vector.

fn asin(self) -> Self

Computes the arcsine of a vector.

fn acos(self) -> Self

Computes the arccosine of a vector.

fn atan(self) -> Self

Computes the arctangent of a vector.

fn atan2(self, x: Self) -> Self

Computes the four quadrant arc-tangent of y(self) and x

fn exp(self) -> Self

The exponential function, returns e^(self)

fn exph(self) -> Self

Half-exponential function, returns 0.5 * e^(self)

fn exp2(self) -> Self

Binary exponential function, returns 2^(self)

fn exp10(self) -> Self

Base-10 exponential function, returns 10^(self)

fn exp_m1(self) -> Self

Exponential function minus one, e^(self) - 1.0, special implementation that is more accurate if the numbr if closer to zero.

fn powf(self, e: Self) -> Self

Computes x^e where x is self and e is a vector of floating-point exponents

fn powiv(self, e: S::Vi32) -> Self

Computes x^e where x is self and e is a vector of integer exponents via repeated squaring

fn powi(self, e: i32) -> Self

Computes x^e where x is self and e is a signed integer

NOTE: Given a constant e, LLVM will happily unroll the inner loop

fn ln(self) -> Self

Computes the natural logarithm of a vector.

fn ln_1p(self) -> Self

Computes ln(1+x) where x is self, more accurately than if operations were performed separately

fn log2(self) -> Self

Computes the base-2 logarithm of a vector

fn log10(self) -> Self

Computes the base-10 logarithm of a vector

fn erf(self) -> Self

Computes the error function for each value in a vector

fn erfinv(self) -> Self

Computes the inverse error function for each value in a vector

fn next_float(self) -> Self

Finds the next representable float moving upwards to positive infinity

fn prev_float(self) -> Self

Finds the previous representable float moving downwards to negative infinity

fn smoothstep(self) -> Self

Calculates a sigmoid-like 3rd-order interpolation function.

NOTE: This function is only valid between 0 and 1, but does not clamp the input to maintain performance where that is not needed. Consider using .saturate() and .scale to ensure the input is within 0 to 1.

fn smootherstep(self) -> Self

Calculates a sigmoid-like 5th-order interpolation function.

NOTE: This function is only valid between 0 and 1, but does not clamp the input to maintain performance where that is not needed. Consider using .saturate() and .scale to ensure the input is within 0 to 1.

fn smootheststep(self) -> Self

Calculates a signmoid-like 7th-order interpolation function.

NOTE: This function is only valid between 0 and 1, but does not clamp the input to maintain performance where that is not needed. Consider using .saturate() and .scale to ensure the input is within 0 to 1.

fn hermite(self, n: u32) -> Self

Computes the n-th degree physicists' Hermite polynomial H_n(x) where x is self and n is an unsigned integer representing the polynomial degree.

This uses the recurrence relation to compute the polynomial iteratively.

NOTE: Given a constant n, LLVM will happily unroll and optimize the inner loop where possible.

fn hermitev(self, n: S::Vu32) -> Self

Computes the n-th degree physicists' Hermite polynomial H_n(x) where x is self and n is a vector of unsigned integers representing the polynomial degree.

The polynomial is calculated independenty per-lane with the given degree in n.

This uses the recurrence relation to compute the polynomial iteratively.

fn jacobi(self, alpha: Self, beta: Self, n: u32) -> Self

Computes the n-th degree Jacobi polynomial via the 3-term recurrence relation.

At the special case where α and β are both zero, the Jacobi polynomial reduces to a Legendre polynomial, and if α and β are both constant zero LLVM can optimize out expressions that rely on those.

NOTE: Given a constant n, LLVM will happily unroll and optimize the inner loop where possible.

Loading content...