Trait thermite::math::SimdVectorizedMath

pub trait SimdVectorizedMath<S: Simd>: SimdVectorizedMathPolicied<S> {
    fn scale(
        self, 
        in_min: Self, 
        in_max: Self, 
        out_min: Self, 
        out_max: Self
    ) -> Self;

    fn lerp(self, a: Self, b: Self) -> Self;

    fn fmod(self, y: Self) -> Self;

    fn hypot(self, y: Self) -> Self;

    fn poly(self, coefficients: &[Self::Element]) -> Self;

    fn poly_rational(
        self, 
        numerator_coefficients: &[Self::Element], 
        denominator_coefficients: &[Self::Element]
    ) -> Self;

    fn poly_f<F>(self, n: usize, f: F) -> Self
    where
        F: FnMut(usize) -> Self;

    fn summation_f<F>(start: isize, end: isize, f: F) -> Result<Self, Self>
    where
        F: FnMut(isize) -> Self;

    fn product_f<F>(start: isize, end: isize, f: F) -> Result<Self, Self>
    where
        F: FnMut(isize) -> Self;

    fn sin(self) -> Self;

    fn cos(self) -> Self;

    fn tan(self) -> Self;

    fn sin_cos(self) -> (Self, Self);

    fn sinh(self) -> Self;

    fn cosh(self) -> Self;

    fn tanh(self) -> Self;

    fn asinh(self) -> Self;

    fn acosh(self) -> Self;

    fn atanh(self) -> Self;

    fn asin(self) -> Self;

    fn acos(self) -> Self;

    fn atan(self) -> Self;

    fn atan2(self, x: Self) -> Self;

    fn exp(self) -> Self;

    fn exph(self) -> Self;

    fn exp2(self) -> Self;

    fn exp10(self) -> Self;

    fn exp_m1(self) -> Self;

    fn cbrt(self) -> Self;

    fn powf(self, e: Self) -> Self;

    fn powiv(self, e: S::Vi32) -> Self;

    fn powi(self, e: i32) -> Self;

    fn ln(self) -> Self;

    fn ln_1p(self) -> Self;

    fn log2(self) -> Self;

    fn log10(self) -> Self;

    fn erf(self) -> Self;

    fn erfinv(self) -> Self;

    fn tgamma(self) -> Self;

    fn lgamma(self) -> Self;

    fn digamma(self) -> Self;

    fn next_float(self) -> Self;

    fn prev_float(self) -> Self;

    fn smoothstep(self) -> Self;

    fn smootherstep(self) -> Self;

    fn smootheststep(self) -> Self;

    fn hermite(self, n: u32) -> Self;

    fn hermitev(self, n: S::Vu32) -> Self;

    fn jacobi(self, alpha: Self, beta: Self, n: u32, m: u32) -> Self;

    fn legendre(self, n: u32, m: u32) -> Self;

    fn bessel_j(self, n: u32) -> Self;

    fn bessel_jf(self, n: Self::Element) -> Self;

    fn bessel_y(self, n: u32) -> Self;

    fn bessel_yf(self, n: Self::Element) -> Self;
}

Set of vectorized special functions optimized for both single and double precision.

This trait is automatically implemented for any vector that implements SimdVectorizedMathPolicied, and delegates to that using the default policy Performance for all methods.

To use a specific execution policy for any function listed below, simply append _p to the function name and provide one of the available policies via turbofish, such as x.sin_p::<Precision>()

Required methods

fn scale(self, in_min: Self, in_max: Self, out_min: Self, out_max: Self) -> Self

Scales values between in_min and in_max, to between out_min and out_max

fn lerp(self, a: Self, b: Self) -> Self

Linearly interpolates between a and b using self

Equivalent to (1 - t) * a + t * b, but uses fused multiply-add operations to improve performance while maintaining precision

fn fmod(self, y: Self) -> Self

Returns the floating-point remainder of self / y (rounded towards zero)

fn hypot(self, y: Self) -> Self

Computes sqrt(x * x + y * y) for each element of the vector, but can be more precise with values around zero.

NOTE: This is not higher-performance than the naive version, only slightly more precise.

fn poly(self, coefficients: &[Self::Element]) -> Self

Computes the sum Σ(coefficients[i] * x^i) from i=0 to coefficients.len()

NOTE: This has the potential to inline and unroll the inner loop for constant input

fn poly_rational(
self,
numerator_coefficients: &[Self::Element],
denominator_coefficients: &[Self::Element]
) -> Self

Computes self.poly(numerator_coefficients) / self.poly(denominator_coefficients) with extra precision if the policy calls for it.

fn poly_f<F>(self, n: usize, f: F) -> Self where
F: FnMut(usize) -> Self,

Computes the sum Σ(f(i)*x^i) from i=0 to n

NOTE: This has the potential to inline and unroll the inner loop for constant input. For best results with dynamic input, try to precompute the input function, as it will allow for better cache utilization and lessen register pressure.

fn summation_f<F>(start: isize, end: isize, f: F) -> Result<Self, Self> where
F: FnMut(isize) -> Self,

Computes Σ[f(n)] from n=start to end.

Returns Ok if the sum converges, Err otherwise with the diverging value.

This will terminate when the sum converges or hits the policy iteration limit.

Note that this method does not accept self, but rather you should use any variables by capturing them in the passed closure.

fn product_f<F>(start: isize, end: isize, f: F) -> Result<Self, Self> where
F: FnMut(isize) -> Self,

Computes Π[f(n)] from n=start to end.

Returns Ok if the product converges, Err otherwise with the diverging value.

This will terminate when the product converges or hits the policy iteration limit.

Note that this method does not accept self, but rather you should use any variables by capturing them in the passed closure.

fn sin(self) -> Self

Computes the sine of a vector.

fn cos(self) -> Self

Computes the cosine of a vector.

fn tan(self) -> Self

Computes the tangent of a vector.

fn sin_cos(self) -> (Self, Self)

Computes both the sine and cosine of a vector together faster than they would be computed separately.

If you need both, use this.

fn sinh(self) -> Self

Computes the hyperbolic-sine of a vector.

fn cosh(self) -> Self

Computes the hyperbolic-cosine of a vector.

fn tanh(self) -> Self

Computes the hyperbolic-tangent of a vector.

fn asinh(self) -> Self

Computes the hyperbolic-arcsine of a vector.

fn acosh(self) -> Self

Computes the hyperbolic-arccosine of a vector.

fn atanh(self) -> Self

Computes the hyperbolic-arctangent of a vector.

fn asin(self) -> Self

Computes the arcsine of a vector.

fn acos(self) -> Self

Computes the arccosine of a vector.

fn atan(self) -> Self

Computes the arctangent of a vector.

fn atan2(self, x: Self) -> Self

Computes the four quadrant arc-tangent of y(self) and x

fn exp(self) -> Self

The exponential function, returns e^(self)

fn exph(self) -> Self

Half-exponential function, returns 0.5 * e^(self)

fn exp2(self) -> Self

Binary exponential function, returns 2^(self)

fn exp10(self) -> Self

Base-10 exponential function, returns 10^(self)

fn exp_m1(self) -> Self

Exponential function minus one, e^(self) - 1.0, special implementation that is more accurate if the numbr if closer to zero.

fn cbrt(self) -> Self

Computes the cubic-root of each lane in a vector.

fn powf(self, e: Self) -> Self

Computes x^e where x is self and e is a vector of floating-point exponents

fn powiv(self, e: S::Vi32) -> Self

Computes x^e where x is self and e is a vector of integer exponents via repeated squaring

fn powi(self, e: i32) -> Self

Computes x^e where x is self and e is a signed integer

NOTE: Given a constant e, LLVM will happily unroll the inner loop

fn ln(self) -> Self

Computes the natural logarithm of a vector.

fn ln_1p(self) -> Self

Computes ln(1+x) where x is self, more accurately than if operations were performed separately

fn log2(self) -> Self

Computes the base-2 logarithm of a vector

fn log10(self) -> Self

Computes the base-10 logarithm of a vector

fn erf(self) -> Self

Computes the error function for each value in a vector.

fn erfinv(self) -> Self

Computes the inverse error function for each value in a vector.

fn tgamma(self) -> Self

Computes the Gamma function (Γ(z)) for any real input, for each value in a vector.

This implementation uses a few different behaviors to ensure the greatest precision where possible.

For non-integer positive inputs, it uses the Lanczos approximation.

For small non-integer negative inputs, it uses the recursive identity Γ(z)=Γ(z+1)/z until z is positive.

For large non-integer negative inputs, it uses the reflection formula -π/(Γ(z)sin(πz)z).

For positive integers, it simply computes the factorial in a tight loop to ensure precision. Lookup tables could not be used with SIMD.

At zero, the result will be positive or negative infinity based on the input sign (signed zero is a thing).

NOTE: The Gamma function is not defined for negative integers.

fn lgamma(self) -> Self

Computes the natural log of the Gamma function (ln(Γ(x))) for any real positive input, for each value in a vector.

fn digamma(self) -> Self

Computes the Digamma function ψ(x), the first derivative of ln(Γ(x)), or ln(Γ(x)) d/dx

fn next_float(self) -> Self

Finds the next representable float moving upwards to positive infinity.

fn prev_float(self) -> Self

Finds the previous representable float moving downwards to negative infinity.

fn smoothstep(self) -> Self

Calculates a sigmoid-like 3rd-order interpolation function.

NOTE: This function is only valid between 0 and 1, but does not clamp the input to maintain performance where that is not needed. Consider using .saturate() and .scale to ensure the input is within 0 to 1.

fn smootherstep(self) -> Self

Calculates a sigmoid-like 5th-order interpolation function.

NOTE: This function is only valid between 0 and 1, but does not clamp the input to maintain performance where that is not needed. Consider using .saturate() and .scale to ensure the input is within 0 to 1.

fn smootheststep(self) -> Self

Calculates a sigmoid-like 7th-order interpolation function.

NOTE: This function is only valid between 0 and 1, but does not clamp the input to maintain performance where that is not needed. Consider using .saturate() and .scale to ensure the input is within 0 to 1.

fn hermite(self, n: u32) -> Self

Computes the n-th degree physicists' Hermite polynomial H_n(x) where x is self and n is an unsigned integer representing the polynomial degree.

This uses the recurrence relation to compute the polynomial iteratively.

NOTE: Given a constant n, LLVM will happily unroll and optimize the inner loop where possible.

fn hermitev(self, n: S::Vu32) -> Self

Computes the n-th degree physicists' Hermite polynomial H_n(x) where x is self and n is a vector of unsigned integers representing the polynomial degree.

The polynomial is calculated independenty per-lane with the given degree in n.

This uses the recurrence relation to compute the polynomial iteratively.

fn jacobi(self, alpha: Self, beta: Self, n: u32, m: u32) -> Self

Computes the m-th derivative of the n-th degree Jacobi polynomial

A the special case where α and β are both zero, the Jacobi polynomial reduces to a Legendre polynomial.

NOTE: Given constant α, β or n, LLVM will happily optimize those away and unroll loops.

fn legendre(self, n: u32, m: u32) -> Self

Computes the m-th associated n-th degree Legendre polynomial, where m=0 signifies the regular n-th degree Legendre polynomial.

If m is odd, the input is only valid between -1 and 1

NOTE: Given constant n and/or m, LLVM will happily unroll and optimize inner loops.

Internally, this is computed with jacobi_d

fn bessel_j(self, n: u32) -> Self

Computes the Bessel function of the first kind J_n(x) with whole integer order n.

NOTE: For n < 2, this uses an efficient rational polynomial approximation.

fn bessel_jf(self, n: Self::Element) -> Self

Computes the Bessel function of the first kind J_n(x) with real order n.

fn bessel_y(self, n: u32) -> Self

Computes the Bessel function of the second kind Y_n(x) with whole integer order n.

NOTE: For n < 2, this uses an efficient rational polynomial approximation.

fn bessel_yf(self, n: Self::Element) -> Self

Computes the Bessel function of the second kind Y_n(x) with real order n.

Loading content...