Trait thermite::SimdBitwise

pub trait SimdBitwise<S: Simd + ?Sized>: SimdVectorBase<S> + Not<Output = Self> + BitAnd<Self, Output = Self> + BitOr<Self, Output = Self> + BitXor<Self, Output = Self> + BitAndAssign<Self> + BitOrAssign<Self> + BitXorAssign<Self> + Shl<S::Vu32, Output = Self> + Shl<u32, Output = Self> + ShlAssign<S::Vu32> + ShlAssign<u32> + Shr<S::Vu32, Output = Self> + Shr<u32, Output = Self> + ShrAssign<S::Vu32> + ShrAssign<u32> {
    const FULL_BITMASK: u16;

    fn bitmask(self) -> u16;

    fn and_not(self, other: Self) -> Self { ... }
}

Defines bitwise operations on vectors

Associated Constants

const FULL_BITMASK: u16

Bitmask corresponding to all lanes of the mask being truthy.

Loading content...
Required methods

fn bitmask(self) -> u16

Returns an integer where each bit corresponds to the binary truthy-ness of each lane from the mask.

Loading content...
Provided methods

fn and_not(self, other: Self) -> Self

Computes !self & other, may be more performant than the naive version

Loading content...