Trait thermite::SimdFloatVector

pub trait SimdFloatVector<S: Simd + ?Sized>: SimdVector<S> + SimdSignedVector<S> {
    type Vi: SimdIntVector<S> + SimdSignedVector<S>;
    type Vu: SimdIntVector<S>;
    fn epsilon() -> Self;

    fn infinity() -> Self;

    fn neg_infinity() -> Self;

    fn neg_zero() -> Self;

    fn nan() -> Self;

    unsafe fn load_f16_unaligned_unchecked(src: *const f16) -> Self;

    unsafe fn store_f16_unaligned_unchecked(&self, dst: *mut f16);

    unsafe fn to_int_fast(self) -> Self::Vi;

    unsafe fn to_uint_fast(self) -> Self::Vu;

    fn sum(self) -> Self::Element;

    fn product(self) -> Self::Element;

    fn mul_add(self, m: Self, a: Self) -> Self;

    fn round(self) -> Self;

    fn ceil(self) -> Self;

    fn floor(self) -> Self;

    fn trunc(self) -> Self;

    fn sqrt(self) -> Self;

    fn is_subnormal(self) -> Mask<S, Self>;

    fn load_f16_unaligned(src: &[f16]) -> Self { ... }

    fn store_f16_unaligned(&self, dst: &mut [f16]) { ... }

    fn combine_sign(self, sign: Self) -> Self { ... }

    fn mul_sub(self, m: Self, s: Self) -> Self { ... }

    fn nmul_add(self, m: Self, a: Self) -> Self { ... }

    fn nmul_sub(self, m: Self, s: Self) -> Self { ... }

    fn fract(self) -> Self { ... }

    fn rsqrt(self) -> Self { ... }

    fn rsqrt_precise(self) -> Self { ... }

    fn recepr(self) -> Self { ... }

    fn approx_eq(self, other: Self, tolerance: Self) -> Mask<S, Self> { ... }

    fn clamp(self, min: Self, max: Self) -> Self { ... }

    fn saturate(self) -> Self { ... }

    fn is_finite(self) -> Mask<S, Self> { ... }

    fn is_infinite(self) -> Mask<S, Self> { ... }

    fn is_normal(self) -> Mask<S, Self> { ... }

    fn is_zero_or_subnormal(self) -> Mask<S, Self> { ... }

    fn is_nan(self) -> Mask<S, Self> { ... }
}

Floating point SIMD vectors

Associated Types

type Vi: SimdIntVector<S> + SimdSignedVector<S>

type Vu: SimdIntVector<S>

Loading content...
Required methods

fn epsilon() -> Self

fn infinity() -> Self

fn neg_infinity() -> Self

fn neg_zero() -> Self

fn nan() -> Self

unsafe fn load_f16_unaligned_unchecked(src: *const f16) -> Self

unsafe fn store_f16_unaligned_unchecked(&self, dst: *mut f16)

unsafe fn to_int_fast(self) -> Self::Vi

Can convert to a signed integer faster than a regular cast, but may not provide correct results above a certain range.

For example, f64 -> i64 is only valid from (-2^51, 2^51) or so.

On instruction sets where this can be done in-hardware, the results should be exact, but do not rely on this for large floats.

unsafe fn to_uint_fast(self) -> Self::Vu

Can convert to a signed integer faster than a regular cast, but may not provide correct results above a certain range.

For example, f64 -> u64 is only valid from [0, 2^52) or so.

On instruction sets where this can be done in-hardware, the results should be exact, but do not rely on this for large floats.

fn sum(self) -> Self::Element

Compute the horizontal sum of all elements

fn product(self) -> Self::Element

Compute the horizontal product of all elements

fn mul_add(self, m: Self, a: Self) -> Self

Fused multiply-add

fn round(self) -> Self

fn ceil(self) -> Self

fn floor(self) -> Self

fn trunc(self) -> Self

fn sqrt(self) -> Self

fn is_subnormal(self) -> Mask<S, Self>

Loading content...
Provided methods

fn load_f16_unaligned(src: &[f16]) -> Self

Load half-precision floats and up-convert them into Self

fn store_f16_unaligned(&self, dst: &mut [f16])

Down-convert self into half-precision and store

fn combine_sign(self, sign: Self) -> Self

Same as self * sign.signum() or select(sign_bit(sign), -self, self), but more efficient where possible.

fn mul_sub(self, m: Self, s: Self) -> Self

Fused multiply-subtract

fn nmul_add(self, m: Self, a: Self) -> Self

Fused negated multiply-add

fn nmul_sub(self, m: Self, s: Self) -> Self

Fused negated multiply-subtract

fn fract(self) -> Self

fn rsqrt(self) -> Self

Compute the approximate reciprocal of the square root (1 / sqrt(x))

fn rsqrt_precise(self) -> Self

A more precise 1 / sqrt(x) variation, which may use faster instructions where possible

fn recepr(self) -> Self

Computes the approximate reciprocal/inverse of each value

fn approx_eq(self, other: Self, tolerance: Self) -> Mask<S, Self>

fn clamp(self, min: Self, max: Self) -> Self

fn saturate(self) -> Self

Clamps self to between 0 and 1

fn is_finite(self) -> Mask<S, Self>

fn is_infinite(self) -> Mask<S, Self>

fn is_normal(self) -> Mask<S, Self>

fn is_zero_or_subnormal(self) -> Mask<S, Self>

fn is_nan(self) -> Mask<S, Self>

Loading content...