Trait thermite::SimdFromCast

pub trait SimdFromCast<S: Simd, FROM>: Sized {
    fn from_cast(from: FROM) -> Self;

    fn from_cast_mask(from: Mask<S, FROM>) -> Mask<S, Self>;
}

Describes casting from one SIMD vector type to another

This should handle extending bits correctly

Required methods

fn from_cast(from: FROM) -> Self

Casts one vector to another, performing proper numeric conversions on each element.

This is equivalent to the as keyword in Rust, but for SIMD vectors.

fn from_cast_mask(from: Mask<S, FROM>) -> Mask<S, Self>

Casts one mask to another, not caring about the value types, but rather expanding or truncating the mask bits as efficiently as possible.

Loading content...