Trait thermite::SimdIntVector

pub trait SimdIntVector<S: Simd + ?Sized>: SimdVector<S> + Eq {
    fn saturating_add(self, _rhs: Self) -> Self;

    fn saturating_sub(self, _rhs: Self) -> Self;

    fn wrapping_sum(self) -> Self::Element;

    fn wrapping_product(self) -> Self::Element;

    fn div_const(self, divisor: Self::Element) -> Self { ... }
}

Integer SIMD vectors

Required methods

fn saturating_add(self, _rhs: Self) -> Self

Saturating addition, will not wrap

fn saturating_sub(self, _rhs: Self) -> Self

Saturating subtraction, will not wrap

fn wrapping_sum(self) -> Self::Element

Sum all lanes together, wrapping the result if it can't fit in T

fn wrapping_product(self) -> Self::Element

Multiply all lanes together, wrapping the result if it can't fit in T

Loading content...
Provided methods

fn div_const(self, divisor: Self::Element) -> Self

For some vectors, this can provide a significant speedup when the divisor is const, as LLVM and Thermite can generate a fixed sequence of instructions to optimally perform the division.

Loading content...